separated setting to different py file now settings loaded from file are in own source file
