+++ tickers/tickers_ETH.txt
@@ -1,89 +1,90 @@
-GRT
-DASH
-EOS
-DATA
-MFT
-XLM
-ICX
-CDT
-IOST
-XVG
-XMR
-STRAX
-WAVES
-COVER
-NAS
-DF
-MANA
-EZ
-IOTA
-ETC
-BRD
-KEY
-VIB
-SCRT
-QSP
-PROS
-ZEC
-DENT
-POWR
-STMX
-BAT
-GLM
-KMD
-QLC
-LTC
-LINK
-QKC
-XEM
-THETA
+ONT
+VET
 GXS
-GHST
-HOT
-ZRX
+IOTA
+HEGIC
+LOOM
+NEO
+IOTX
+XMR
+GRT
+AION
+QLC
+XEM
+BLZ
+BNT
+POWR
+XLM
+GLM
+MANA
+KMD
+IOST
+SCRT
+SNT
+PROS
+EOS
+ADA
+ETC
+FIRO
+XVG
+CDT
 CVC
-MTL
-KNC
-OST
+COVER
 DEXE
-PIVX
+QTUM
+SLP
+RLC
+SC
+BETH
+HOT
+EZ
+LSK
+KNC
+ZIL
+OST
 UFT
 LRC
-BQX
-ADX
-WAN
-BNT
-ENJ
-VET
-BLZ
-CVP
-SLP
-ADA
-AION
+DENT
+PIVX
 STEEM
-XRP
-RLC
-ELF
-QTUM
-NANO
-FIRO
-BNB
-NEO
-HEGIC
-LSK
+MTL
+QKC
+WBTC
+THETA
+WAVES
+NEBL
+DATA
+OMG
+CVP
+TRX
+REP
+QSP
+MFT
+VIB
+STRAX
+ICX
+DF
+DASH
 ZEN
 AAVE
-OMG
-ZIL
-TRX
-FUN
-SNT
-WBTC
-SC
-NAV
-NEBL
+ADX
+ENJ
+BAT
+LTC
+BNB
 PUNDIX
-IOTX
-REP
-LOOM
-ONT
+WAN
+NAV
+BRD
+ZRX
+XRP
+KEY
+NAS
+ZEC
+STMX
+NANO
+GHST
+FUN
+BQX
+ELF
+LINK
